Create Wikipedia_Node_Edge_Extractor.py
Script for collecting Wikipedia page data as well as internal hyperlinks (Wikipedia pages that point at each other) and storing that data as node and edge lists. These lists can be used to visualize Wikipedia page as a network. Note this script must be seeded with a page or list of pages you want to review first.